requirements.txt
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-aiml==0.9.1
Werkzeug==0.15.4
WTForms==2.2.1
